Theme Showcase: Extract ThemeSection, add dark variant, fix section padding
Extract ThemeSection into its own file for reusability. Add dark
variant support with CSS custom properties for the Partner section.
Move horizontal padding from the showcase wrapper to individual
sections so the dark section background extends edge-to-edge.
