+++ client/my-sites/themes/banners-modern/style.scss
@@ -34,7 +34,7 @@
 .banner-modern__description {
 	color: var( --studio-gray-80 );
 	font-size: $font-body-large;
-	line-height: 1.2;
+	line-height: 1.5;
 	margin-block-end: 24px;
 }
