+++ client/my-sites/themes/theme-showcase.scss
@@ -635,7 +635,7 @@
 // Modern theme card overrides (logged-out redesign)
 .theme-showcase.is-modern {
 	.themes__showcase {
-		padding: 0 24px;
+		padding: 0;
 	}
 
 	.themes-list {
